cleanup: Resolve some of dependencies on private headers
