+++ include/iocore/net/EventIO.h
@@ -22,7 +22,10 @@
  */
 
 #pragma once
-#include "../../../src/iocore/net/P_UnixPollDescriptor.h"
+
+#include "tscore/ink_config.h"
+
+struct PollDescriptor;
 
 using EventLoop = PollDescriptor *;
 
@@ -51,6 +54,12 @@ using EventLoop = PollDescriptor *;
 #else
 #define INK_EV_EDGE_TRIGGER 0
 #endif
+#include <sys/event.h>
+#define INK_EVP_IN    0x001
+#define INK_EVP_PRI   0x002
+#define INK_EVP_OUT   0x004
+#define INK_EVP_ERR   0x010
+#define INK_EVP_HUP   0x020
 #define EVENTIO_READ  INK_EVP_IN
 #define EVENTIO_WRITE INK_EVP_OUT
 #define EVENTIO_ERROR (0x010 | 0x002 | 0x020) // ERR PRI HUP
