+++ include/iocore/net/quic/QUICStream.h
@@ -25,7 +25,6 @@
 
 #include "tscore/List.h"
 
-#include "../../../../src/iocore/eventsystem/P_VConnection.h"
 #include "iocore/eventsystem/Event.h"
 
 #include "iocore/net/quic/QUICConnection.h"
