+++ include/proxy/FetchSM.h
@@ -29,7 +29,6 @@
 
 #pragma once
 
-#include "../../src/iocore/net/P_Net.h"
 #include "proxy/http/HttpSM.h"
 #include "proxy/http/HttpTunnel.h"
