+++ src/iocore/net/SSLStats.cc
@@ -25,6 +25,7 @@
 
 #include <openssl/err.h>
 
+#include "iocore/net/SSLMultiCertConfigLoader.h"
 #include "P_SSLConfig.h"
 #include "P_SSLUtils.h"
 #include "../../records/P_RecProcess.h"
